[PCP-9130] add securityContext to the pipeline
