+++ charts/generic-runner/Chart.yaml
@@ -6,7 +6,7 @@
 
 apiVersion: v2
 name: generic-runner
-version: "1.0.6"
+version: "1.0.7"
 appVersion: "1.0.0"
 description: generic-runner
 type: application
