+++ charts/helm-install/Chart.yaml
@@ -6,7 +6,7 @@
 
 apiVersion: v2
 name: helm-install
-version: "1.0.10"
+version: "1.0.11"
 appVersion: "1.0.0"
 description: helm-install
 type: application
