+++ charts/helm-install/templates/task.yaml
@@ -33,6 +33,8 @@ spec:
     - name: {{ .Values.name }}
       image: {{ .Values.pipelineImage }}
       workingDir: /workspace
+      securityContext:
+        {{- toYaml .Values.securityContext | nindent 8 }}
       timeout: "{{ .Values.taskRunTimeout }}"
       volumeMounts:
         - name: {{ .Values.commonScriptConfigmapName }}
