+++ charts/helm-install/values.yaml
@@ -42,3 +42,28 @@ taskRunTimeout: 1h55m0s
 finallyTimout: 5m
 # the service account that this pipeline will be attached to
 serviceAccount: provision
+
+# the container and tools need to be ready to run as non-root user.
+# follow Restricted: https://kubernetes.io/docs/concepts/security/pod-security-standards/
+podSecurityContext:
+  seccompProfile:
+    type: RuntimeDefault  # RuntimeDefault or Localhost (must for Restricted)
+  runAsNonRoot: false # true  # Prevents processes from running as root (both) Container must run as a non-root user (must for Restricted)
+  runAsUser: 0 # 1000  # Explicitly set the user ID for running processes (both) (must for restrict, non-zero or undefined/null)
+  # runAsGroup: 3000  # Set the group ID for running processes (both)
+  # fsGroup: 2000  # Specify the group ID for mounted volumes (pod security context only)
+  # fsGroupChangePolicy: OnRootMismatch # Set the policy for the change of fsGroup in the volume (pod security context only)
+  # supplementalGroups: [1000,3000]  # Specify the group IDs for running processes (pod security context only)
+  # sysctls: {} # Kernel settings to apply (pod security context only)
+  # seLinuxOptions: {} # SELinux options to apply (both)
+  # windowsOptions: {} # Windows options to apply (both)
+
+# the container and tools need to be ready to run as read-only root filesystem.
+securityContext:
+  readOnlyRootFilesystem: false # true  # Prevents the container from writing to the root filesystem (container security context only)
+  allowPrivilegeEscalation: false  # Prevents privilege escalation (container security context only) (must for Restricted)
+  capabilities:  # Add or drop capabilities (container security context only)
+    drop: ["ALL"]  # Drop all capabilities (must for Restricted)
+    # add: ["NET_BIND_SERVICE"] # Add the NET_BIND_SERVICE capability to the container (only for lower 1024 ports) (allow for Restricted)
+  # privileged: false # Prevents privilege escalation (container security context only) (for Baseline)
+  # procMount: Default # Set the mount propagation mode for the container (container security context only)
