+++ charts/provisioner-config-local/recipes/tp-automation-o11y.yaml
@@ -54,6 +54,8 @@ meta:
     # flow control
     GUI_TP_PORT_FORWARD: true
     GUI_TP_OVERWRITE_DNS: false
+    GUI_TP_ACTIVE_USER: true
+    GUI_TP_CREATE_DP: true
   globalEnvVariable:
     REPLACE_RECIPE: true
     PIPELINE_CHECK_DOCKER_STATUS: false
@@ -65,6 +67,7 @@ meta:
     # automation setup
     TP_AUTO_CP_VERSION: ${GUI_TP_AUTO_CP_VERSION:-"1.3"}
     TP_AUTO_IS_CONFIG_O11Y: ${GUI_TP_AUTO_IS_CONFIG_O11Y:-false}
+    TP_AUTO_SCRIPT_FOLDER: /tmp/auto
 
     # CP setup
     TP_AUTO_CP_INSTANCE_ID: ${GUI_TP_AUTO_CP_INSTANCE_ID:-"cp1"}
@@ -101,9 +104,12 @@ meta:
 
     # flow control
     PYTHON_FILE_INPUT_NAME: py-scripts.yaml
+    PYTHON_FILE_LOGIN_POINT: page_auth.py
     PYTHON_FILE_ENTRY_POINT: run.py
     TP_PORT_FORWARD: ${GUI_TP_PORT_FORWARD:-true}
     TP_OVERWRITE_DNS: ${GUI_TP_OVERWRITE_DNS:-false}
+    TP_ACTIVE_USER: ${GUI_TP_ACTIVE_USER:-true}
+    TP_CREATE_DP: ${GUI_TP_CREATE_DP:-true}
 tasks:
 - condition: ${TP_OVERWRITE_DNS} # this is used for overwrite DNS
   clusters:
@@ -137,8 +143,9 @@ tasks:
           echo "Created file: $file_name"
         fi
       done
+      mkdir -p ${TP_AUTO_SCRIPT_FOLDER}
+      cp * ${TP_AUTO_SCRIPT_FOLDER}
       pip install -r requirements.txt
-      python ${PYTHON_FILE_ENTRY_POINT}
   payload:
     base64Encoded: false
     fileName: ${PYTHON_FILE_INPUT_NAME}
@@ -148,3 +155,21 @@ tasks:
       {{ base $path }}: |
       {{ $root.Files.Get $path | b64enc | indent 2 -}}
       {{- end }}
+- condition: ${TP_ACTIVE_USER}
+  clusters:
+    - name: ${TP_CLUSTER_NAME}
+  script:
+    ignoreErrors: false
+    fileName: script.sh
+    content: |
+      cd ${TP_AUTO_SCRIPT_FOLDER}
+      python ${PYTHON_FILE_LOGIN_POINT}
+- condition: ${TP_CREATE_DP}
+  clusters:
+    - name: ${TP_CLUSTER_NAME}
+  script:
+    ignoreErrors: false
+    fileName: script.sh
+    content: |
+      cd ${TP_AUTO_SCRIPT_FOLDER}
+      python ${PYTHON_FILE_ENTRY_POINT}
